+++ wssession/outbound.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	gtkitjson "github.com/gtkit/json"
+	gtkitjson "github.com/gtkit/json/v2"
 
 	"github.com/gorilla/websocket"
 )